@@ -12,7 +12,8 @@
         },
         "allow-plugins": {
             "composer/package-versions-deprecated": true,
-            "simplesamlphp/composer-module-installer": true
+            "simplesamlphp/composer-module-installer": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true
         }
     },
     "autoload": {
@@ -26,10 +27,10 @@
         }
     },
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
 
         "simplesamlphp/composer-module-installer": "^1.3.4",
-        "simplesamlphp/simplesamlphp": "^3@dev",
+        "simplesamlphp/simplesamlphp": "^2.1",
         "simplesamlphp/xml-cas": "^1.0.3",
         "simplesamlphp/xml-common": "^1.12.2",
         "simplesamlphp/xml-soap": "^1.3.0"
